Moved RawFileXRDData to schema.py and added alias
